requirements.txt
pytorch
pandas
scikit-learn
ctgan
opacus